requirements.txt
django==1.8.4
djangorestframework==3.1.3
django-filter==0.10.0
Markdown==2.6.2
requests==2.9.1
pytest==2.8.7
bokeh==0.11.1
